Fix not initialized pointer when parser is starting on a new file.
The pointer oldEntry holds the last value of VhdlParser::current and is
compared to detect comment blocks that belong together. This could lead to
false positives when a new file is parsed and the new entry in
VhdlParser::current has by change the same pointer value as the old entry.
Doxygen warning like "warning: Found unknown command `\brief'" are the
consequence that leads to concatenated brief and detailed description in
the resulting output.