Make it possible to list namespace members in file scope for XML output.
For better consistency with the HTML output, where each file
documentation lists (and links to) all members of given namespace. This
also makes it possible to be consistent with the HTML output in case
a namespace is not documented and thus all its member detailed docs
should be put into corresponding file docs instead.

In order to be backwards compatible and avoid breaking stuff for
existing users of the XML output, this is controlled by a new
XML_NAMESPACE_MEMBERS_IN_FILE_SCOPE configuration option that defaults
to NO.

Note that this, unlike the HTML output, will put the whole detailed docs
into the file scope instead of just listing them. It's up to the user of
the XML output to deduplicate this information. It can be done for
example by comparing member ID prefixes with compound ID -- íf
different, the detailed docs are already somewhere else.