Create test possibilities for xhtml and pdf output
The tests in the testing directory provided possibilities to do some regression tests based on the xml output.
With the option --xhtml it is possible to see whether or not a test or set of tests is xhtml compliant.
With the option --pdf it is possible to see whether or not a test or set of tests can be build to a pdf file (per test).