Problem with TEST_FLAGS when using cmake for Ninja
The command sequence:
    cmake -G Ninja -D CMAKE_BUILD_TYPE=Release ..
    cmake --build . -- -j 3 -v || echo error && cd .. && exit 1
results in the error message:
    ninja: error: build.ninja:2384: bad $-escape (literal $ must be written as $$)

In respect to LEX_FLAGS and YACC_FLAGS (the later still required some quotes), similar problems appeared in the past and this was solved by setting the (local) LEX_FLAGS and YACC_FLAGS to the empty string. A similar solution didn't work directly for the TEST_FLAGS so here the code has been duplicated once with the flags and once (for Ninja) without the flags.