Making the 'tex' part of \makeindex available to the user
In case we need to use another 'makeindex' command in the Makefile / make.bat we can use the configuration tag MAKEINDEX_CMD_NAME
When we want to have another index we can use e.g. EXTRA_PACKAGES = [nottoc]tocbibind
but in those cases  the \makeindex command is still the same but should be \makeindex[intoc]. By means of the new configuration tag LATEX_MAKEINDEX_CMD this discrepancy has been solved.
Due to the default value some small changes in the configuration parser were necessary as well.

(based on the stack question https://stackoverflow.com/questions/44394311/add-index-to-toc-with-doxygen).