Bug 663297 - @todo in @param leads to strange confusing message
With the 1.8.14 version the message was:
  warning: unexpected token in comment block while parsing the argument of command param
this is due to the fact that the @todo command is replaced by a \xrefitem command but it was joined directly with the parameter name, adding a space, in the code, solves the problem.
In case there was another word (or doxygen command) in front of it the problem does not occur as it \xrefitem is not joined with the parameter name.
Similar problem occured with the \endinternal (\internal) command.