Bug 782754 - rtf generation
Looks like problem has been introduced with the fix for bug 445105 (release 1.5.3), in this case the \par is necessary.
The fix for bug 741547 (pull request #596) is related to this problem and the fix here is not necessary with this fix.
Related to bug 778525, here 2 problems occur and the mangling is solved with this patch.