Bug 796494 - Collapsed treeview arrow displays as emoji in Microsoft Edge
The right  arrow (&#9654;) is inn MicrosoftEdge displayed as an Emoji it displays as a white arrow inside a blue square with a black border instead of a monochrome arrow. Unicode contains another right arrow (&#9258) which displays correctly (also in Chrome, Opera, FireFox, Internet Explorer; in some even better).

The Unicode naming is a bit confusing:
- for the down arrow we only have:
  - &#9660;  'BLACK DOWN-POINTING TRIANGLE'
- for the right arrow we have:
  - &#9654; 'BLACK RIGHT-POINTING TRIANGLE'
  - &#9658; 'BLACK RIGHT-POINTING POINTER'
so it looks logical to use both Triangles but this is not the case.