Bug 363762 - Tooltips are not shown in dot-generated graphs
The basic problem mentioned in this issue was already solved in Mai 2009 (version 1.5.8 or earlier), in case a brief description is present this is displayed otherwise the node  number is specified)
A remaining thing was that the tooltip with the root nodes (gray nodes) was always something like Node0 irrespective of the brief description. This has been solved here.