redundant input_filter runs significantly reduce performance when FILTER_SOURCE_FILES and INLINE_SOURCES are both enabled
When filtering output write the result to a temporary file that can be reread in case  the next call to the filter routine is again for the same file  (e.g. for inline source code or code snippets). Reading a file is much faster than running an popen call i.e. spawning a subprocess  (that would anyway have to read the file and  and the result has to be read from a pipe).