Bug 748927 - Navigation incorrect with escaped symbols
Due to the fact that constructs like `<my>` in a page title are seen as an XML-tag (with a resulting warning) the `<` should be escaped but this leads to the fact that the escape sign is shown in the bars on top of a HTML page.
The basic problem is due to the fact that page titles are not really interpreted by doxygen (as "nothing" can be, generically, handled in the title of a page.