Update doxygen.cpp
Reading a few thousands of input CPP/H files in alphabetical order is very slow because of using the inSort construct. The inSort function does a linear loop through the list to find the right alphabetical position, leading to a quadratic algorithm. During the creation of the file list there is no need to keep it sorted - it is not used at all yet. Therefore I propose to just append them and to a final sort at the end of the creation of the file list.
For my use case with several thousands of input files, this reduces creating the file list from about an hour to a few seconds !!