declares XMLCodeGenerator in xmlgen.h
In order for sqlite3gen to re-use parts of the XML generator, XMLCodeGenerator
needs to be declared in a header file. I parroted how other generators handle
this and have it working in both the XML and sqlite3 outputs, but I don't have
any sense of whether this is otherwise "right".