Handling of mainpage and its subpages
The mainpage has a bit of another internal structure in doxygen than other pages and has to be handled as such. Furthermore in docbook output pages (chapters, sections, ...) should be closed when all information regarding the page is in.
(The problems here were found based on " Pull Request #6542 Bug 751700 - Main page absent in TOC of CHM, if PROJECT_NAME is empty", issue #5818 / #6542)
- latexgen, rtfgen see to it thatsu pages of a mainpage are seen as subpages and not as new chapters
- index, pagedef, groupdef new sub items as some output types (docbook, others won't do anything here) need to be closed at another moment.
- docbookgen handling of pages, seeing to it that titles are at the right place and that the opening and closing  is done properly
- doxygen, docbookgen in docbook the mainpage has its own file, but this does not work properly (double levels of chapter), so the mainpage file needs to be included into the index file