Flexibility in setting Python protection
Python does not have a real knowledge of protection in respect to variables and functions (i.e. public / private / protected). There are some ad-hoc rules but they changed over time as well.
Doxygen uses  some build in settings for protection i.e. methods starting with '__' and ending with '__' are public other variables and methods starting with '_' are private. This change makes it possible for the user to define his own rules for the protection within Python.

(Loosely based on issue #6554)