issue_6594 using ingroup and anchor causes tests to disappear
The determination whether or not a reference item was in a list was based on the list name and it returned the first list item id. With multiple items only the first item was found.
Checking whether an item should be appended is now based on the list name and the item id.