Adding perlmod tree nodes  documentation to LaTeX / PDF
Adding the perlmod tree nodes documentation in an identical way to HTML and LaTeX / PDF.
A requirement for this is that LaTeX supports more than the default 4 levels in the 'itemize' environment, therefore the 'itemize' environment has been extended to 12 levels (and for consistency purposes also the 'enumerate' environment). This is all handled in the 'doxygen.sty' file.