Wrong counting of lines during VHDL code output
The code coloring of the output stops at an early stage, resulting in incorrect HTML / LaTeX / ... output, due to a double counting of the lines in case of `STRIP_CODE_COMMENTS=NO`.

Problem is visible when looking at the documentation of the VHDL documentation block in the manual for VHDL code in e.g. HTML and LaTex (Note problem occurs in the 1.8.15 master, not in the 1.8.14 version).