XHTML incorrect attribute values for align and valign
The attribute values of the `align` and `valign` attribute have to be in lowercase and the `valign` attribute value `center` does not exist but has to be `middle`.
Most browsers do accept the 'incorrect' values but it is better to have the right values present.
(Found by means of the CGAL/cgal repository)