XHTML problem with name attribute with VHDL name attribute
When running xhtml checker on the doxygen diagram example we get:
Syntax of value for attribute name of a is not valid
Document mux/xhtml/classmux__using__with.xhtml does not validate
This is due to a space in the name tag, substituting the appropriate code solves the problem.
As this is a VHDL specific problem only these strings are converted.