Make var & function declaration/definition location not order-dependent.
Previously, if a definition was processed before declaration, both the
declaration and definition location was set to the definition. It was
working correctly only when declaration was found before definition, but
due to the fact that C/C++ files are named like this:

    foo.cpp
    foo.h

And thus whe globbing the filesystem, the definition is always seen
before the declaration. Note that this affects only functions inside
namespaces, (in my testing) functions in global scope were reported
twice, once for each file (which can be considered okay, as it doesn't
exhibit this bug).

For variables, I'm distinguishing definitions and declarations based on
presence of the extern keyword -- declarations have it, definitions not.

Note that this is still a bit random when there's more than one
var/function declaration, but the new behavior is still better than
conflating the definition/declaration locations into one. Also it's
very probable that only one declaration will have a corresponding
Doxygen doc block and others can be hidden from Doxygen using a
preprocessor.