Implement a new EXTRACT_PRIVATE_VIRTUAL option.
The classic article about virtuality from Herb Sutter [1] suggests that
virtual functions are always private and public class interface is never
virtual. Until now, it was not really possible to document these
functions in Doxygen:

 * Enabling EXTRACT_PRIVATE would show all internals, not just virtual
   functions, which is not wanted.
 * Enabling HIDE_UNDOC_MEMBERS and HIDE_UNDOC_CLASSES would effectively
   disable warnings about *all* undocumented members, which is not wanted.

The usual workaround was to put the members into protected scope just
for Doxygen:

    #ifdef DOXYGEN_GENERATING_OUTPUT
    protected:
    #else
    private:
    #endif
        /** @brief Documented private virtual function */
        virtual doStuff();

The new EXTRACT_PRIVATE_VIRTUAL option makes these visible (and able to
be linked to), but shows them *only* if they are documented.

[1] http://www.gotw.ca/publications/mill18.htm