Terminating brief command
The brief command is intended for a small description of a function / class / namespace etc. and not to have extensive formulas or code snippets. These commands should terminate the brief description and start the detailed section.
The commands `\snippet` and `\snippetlineno` are added to the list of commands to terminate the brief description.
The commands `\f[` and `\f{` are handled separately but now also start a detailed description.

The `\f$` is just for small formulas and thus possible in a brief description.