issue #6679 Multiple use of section label warning for Setex-stype headers in markdown
Due to the move of the markdown code to another place the adding of the section info of section, subsection etc. would be attempted to add twice (first in markdown and later on again when translating the resulting '`section`, `\subsection` commands. The adding should only take place when handling the (replaced) commands.

Removing the test in `addSection` won't work as in that case using twice a sectioning command with the same label would not be signaled anymore.