issue #6791 TOC not generated when using a particular Markdown header style
The default for the `TOC_INCLUDE_HEADERINGS` was set to 0, for compatibility this should be 5.
(Problem was observed in the doxygen documentation, chapter markdown, where the TOC was missing; alternatively in the Doxyfile the `TOC_INCLUDE_HEADERINGS` could be set to `5`,. The solution chosen is more consistent with the old situation).