Bug 757574 - Warning regarding subsection with anchor in markdown
In case `##` sections are used with and label marker and  the previous level had no label marker the previous level was rendered as `<h...>`, now they are rendered as doxygen sections and a label is generated. In case of a tag file the generated label is excluded as otherwise the same generated  label could exist in different tag files and give warnings.