Section title runs into margin
In the PDF version of the doxygen manual we see that for the `\image` command the chapter title the text runs into the margin (due to an unbreakable part).
With this fix the problem is solved.
(see also: https://tex.stackexchange.com/questions/474055/section-title-running-into-margin and , for the ifstar part, https://tex.stackexchange.com/questions/376375/using-ifstar-to-define-a-star-variant)