Incorrect layout of class index (HTML)
Regression of #6663
When looking at the example of #4935 the sorting and layout is correct in 1.8.14 but the layout is incorrect in 1.8.15.
The incorrect layout is due to #6663 where a special case was solved for XHTML, the correct procedure should have been that in case of a table row without cells a dummy cell is added.
The changes from #6663 have been reverted and the fix has been corrected (the special case occurred in the doxygen diagrams example).