Latex code double minus goes to endash
In case in code 2 consecutive minus signs are present they are shown as endash.
using the same code as in the non code part solves the issue.
Issue can e.g. be seen in cpp source code (with e.g. i--) and in the doxygen documentation in the paragraph "Comment blocks in VHDL"
This is a regression in 1.8.15