Problem with referenced / references relations
Main problem is that when the documentation is in the `.h` file, including e.g. `\showrefby` this is overwritten by the global setting from the `.cpp` file.
Strategy is that in case something is explicitly set this should be used (resulting in a function with both fields, i.e. from `.h` and `.cpp`) when merging the documentation  (i.e. `addMemberDocs` in doxygen.cpp).
On other places where a `root` and member are merged the `||` operation has also been changed to  the 2 agruent version.