Show git version information
Show the git version on designated places (currently version, help and Doxyfile difference).
The implementation is based on the git_watcher.cmake (https://github.com/andrew-hardin/cmake-git-version-tracking).
The information is useful to see which. master, version of doxygen is used for a build on systems where a lot of builds are make (Fossies.org) or very regular builds are made (CGAL)

Furthermore the tracking of the VERSION file was not done (when changed the version.cpp was not updated), this is corrected as well.