URL contains $ sign on second run in map of images
When running doxygen twice on a repository the generation of e.g. 'included directory' images  (also seen for a.o. 'class' image) is not done again and the available map file is used, but the replacement of some terms is not done properly (the $ at the beginning is not removed)