Ambiguous file name in file command
I some cases a file name is used with the `\file` command and this filename exists multiple times although the `\file` command is in the file itself (actually file name can be omitted).
Another possibility is that the file description is in another file in the same directory.
When file is not found or ambiguous, try to prepend the path of the current file to it and see if the file can be found.

Example is shown in the warnings file like:
```
gmic-2.6.1/gmic-qt/src/Common.h:3: warning: the name `Common.h' supplied as the second argument in the \file statement matches the following input files:
   gmic-2.6.1/gmic-qt/src/Common.h
   gmic-2.6.1/zart/include/Common.h
Please use a more specific name by including a (larger) part of the path!
```
and here it is quite clear and unambiguous which file is meant.