Incorrect Reached end of file while still inside a (nested) comment for TCL / Python
In case we have in the comment (or in the code) of a TCL file of Python file a construct like:
```
proc get_suite {dir {sort 1}} {
 set files [glob -nocomplain $dir/*.bin]
 set files [glob -nocomplain $dir/*.ps]
}
```
we get the warning:
```
warning: Reached end of file while still inside a (nested) comment. Nesting level 2 (probable line reference: 3, 2)
```
although the '/*' construct has no special comment meaning in TCL / Python (comment signs '#' / '#' or '"""). So if a c-comment construct is found it is ignored for TCL and Python.