Special handling of the UTF8 nbsp sequence
The UTF8 nbsp sequence 0xc2 0xa0 is not seen as a whitespace sequence and not handled properly.
This can lead to:
```
warning: found </c> tag without matching <c>
```
when we have e.g. ```e.g. `linux`<br>``` where the space between the `.` and the backtick is actually the UTF8 nbsp sequence
Replacing the the UTF8 nbsp sequence with the `&nbsp;` sequence.