More than 26 appendices in LaTeX manual
The 1.8.15 manual had 26 examples but due to the fact that a new example was created we got 27 and thus the characters A-Z are not sufficient.
In a number of LaTeX versions this leads to messages like:
```
Appendix \GenericError {               }{LaTeX Error: Counter too large}{See th
e LaTeX manual or LaTeX Companion for explanation.}{You've lost some text.  Try
 typing  <return>  to proceed.\MessageBreak If that doesn't work, type  X <retu
rn>  to quit.}.

! LaTeX Error: Counter too large.
```

Solution for this is to reorganize the examples a little bit. This is done here by means of the "groups":
 - Style examples
 - Language Examples

as a consequence also the sections / subsections etc had to go 1 level deeper and this is accomplished by defining a new environment in which the different sections go 1 level deeper.