issue #6925 Missing ATX headings in markdown pages
In general ATX headers ('#' headers and other markdown ) are converted to doxygen `\section` etc. commands
In case not all levels of doxygen commands are present so like;
```
\section  sect1 Section
....
\subsubsection subsubsect1 Sub sub section

```
the information of the `\subsubsection` is not shown. Same happens in case the `\section` is not present at all.
Part of the steering of the ATX headers is done by means of the configuration setting TOC_INCLUDE_HEADINGS (default is 5) setting this setting to '0' will not convert the '#' headers to `\section` etc. (but this as side note).

Basic problem is fixing the not shown headers in case of missing levels in the `\section` commands.
- In case the `\section` command does start the page
- In case the `\section` command does not start the page.

Case the `\section` command does start the page: this is handled in the routine `int DocSection::parse()`
- see to it that not only the exact matching level is handled but also the 'jumps' in levels
- loop all 'section' of the found type and its sub sections and only jump out in case of the 'jump' level case
- give a warning in case of a 'level jump'

Case the `\section` command does not start the page: this is handled in the routine ` void DocRoot::parse()`
- as e.g. `\paragraph` could be followed by a `\subsubsection` each section type has to be handled one after another (otherwise higher types would not be handled)
- note the order is important therefore a `\subsubsection` has to be handled after a `\paragraph.
- due to the different handling the 'TK_LISTITEM' has to be handled at the end (each part can also give a 'TK_LISTITEM'