Show git version information
The original version has as features:
- getting the git version number for usage in doxygen
- making the doxygen version number inclusion dependent on the VERSION file

The disadvantage of the chosen methodology was that an extra correction step was necessary, by defining getter methods to retrieve the values this correction can be hidden.
The information is coming from different sources:
- the VERSION file
- the git "repository and build system (when present)

Furthermore there are a couple of places where the version information is used (a.o. doxygen and doxywizard executable, though the doxygenwizard was only done "half hearted")

The handling of the VERSION file has been made in such a way that it is comparable with the generation of the git version changes.

For a better abstraction the version handling is all done in a separate directory.