Correction warning message
When there is a problem with a backticjk comment block we can get the message:

reached end of file while inside a ````' block!

this is a bit unreadable / non understandable as we see see 4 backticks followed by a single quote, it woulb be more readable when we get

reached end of file while inside a '```' block!