End comment marker in \code section.
When we have an end comment in a `\code` segment in a section we can get a message like:
```
warning: found */ without matching /* while inside a \code block! Perhaps a missing \endcode?
```
in the similar situation in a `\verbatim` section this message is not given. Switching from `\code` to `\verbatim` is not an option here as in that case other "commands" can lead to problems like in a code  like `$string =~ s/^\s*//;`

In this patch the start `/*` and end `*/` comment markers for `\code` and `\verbatim` are completely ignored.