Problems with escaped tilde
In some input formats it is necessary to escape a tilde (~) and this leads to a conflict with the selection of the language specific filter in doxygen.
things that lead to problems are e.g. `~user` or `~1MBit`.

This patch checks whether the used 'LanguageId' really is a doxygen supported language (note it is possible that the language is not enabled in the used doxygen executable).