Don't remove `*` in case of `*/` inside code block (at start of line)
When having in a `\code` block  a comment line like ` * statement` the `*` is removed as it is (probably) just required from the coding style / readability in the code.
When having though a line `*/` in a `\code` block the `*` is also removed and in this case it should not be don.