Missing brief descriptions with `\defgroup`
Due to the fact that issue #6945 ended a brief description, all following was put into the detailed description (unless explicit \brief was used). The `\defgroup` should end the current brief description, but first line of the `\defgroup` should be seen as a new brief description.
Problem can be seen with in the doxygen documentation in the example in the Grouping part and also with #7053 a small example is given in the comment.