Consistency of configuration parameters
Configuration parameters were used 2 forms in the code bare name like `PAPER_TYPE` and as string `"PAPER_TYPE"`.
In case of a non existing "bare name" (or misspelling) this is signaled by the compiler but in case of the string this was only shown at runtime.
In case of the usage of an obsolete configuration parameter the "bare name" was signaled as not existing (anymore).

The essential change here is the place where the ConfigValues are initialized.