Problem with '<td nowrap>'
In the docbook ouput this was shown with the attribute `nowrap>=''`, in HTML this empty tag was skipped.
Normally a HTML attribute will have a value but in some cases it is possible without attribute and when this is the last attribute the `>` was accidently added to the attribute (in case of the value the `>` was already considered).
Furthermore `In XHTML, attribute minimization is forbidden, and the nowrap attribute must be defined as <td nowrap="nowrap">.`, this is now handled for HTML and docbook as well.