Opening and ending tag mismatch: tbody in docbook output
In case of docbook output and having nesting tables (e.g. in the tables chapter of the doxygen documentation), we get invalid xml for docbook as the `<tbody>` is not closed.
Each table level should have an own flag to signal whether or not the `<tbiody>` tag is set or not, otherwise the inner table resets the outer tables its "flag".