Incorrect NCName in docbook citations
In a bibtex reference it is possible to have a colon as the id, but in docbook this will result in:
```
docbook/citelist.xml:5: validity error : xml:id : attribute value _citelist_1CITEREF_pre:post is not an NCName
<varlistentry><term><anchor xml:id="_citelist_1CITEREF_pre:post"/>[1]</term>
```

So the id has to be translated. A colon in an id can only happen in case the names is provided from the outside.