issue #7190 1.8.16: Blank FILE_PATTERNS => no files processed
This is actually a regression on #7195. Due to the fact that the init() routine has moved up, the settings for `FILE_PATTERNS` have already been done and a call to init does not operate on the `FILE_PATTERNS` variable but on the structure from which the `FILE_PATTERNS` was derived. We have should here operate on the `FILE_PATTERNS` directly as well., bu getting the default values out of the underlying structure.