Protection of interfaces and types in Fortran
In the proposed pull request #7192 an attempt was made to get the protection of the types and interfaces correct. This looks like to work in case the types end interfaces are on continuation lines, but as soon as all types and interfaces are in one line this failed (see discussion with #7192).

in this patch it is solved by updating the so called modifiers also for CLASS_SEC (types and interfaces are seen as "classes"), furthermore in case the modifiers are not found a test is done whether or not the module (scope) name is prepended to the entry name as with the modifiers this module name is not present.